GCP/02_Code/01_CloudFunction/Transcribe/requirements.txt
google-cloud-speech
google-cloud-storage
google-cloud-firestore
soundfile
numpy
